create changeset and run api-reports
