+++ .changeset/young-olives-cheer.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-github-issues': minor
+---
+
+Add filtering and ordering to the graphql query
