+++ packages/app/src/components/catalog/EntityPage.tsx
@@ -153,8 +153,6 @@ import {
   ReportIssue,
 } from '@backstage/plugin-techdocs-module-addons-contrib';
 
-import { GitHubIssuesCard } from '@backstage/plugin-github-issues';
-
 const customEntityFilterKind = ['Component', 'API', 'System'];
 
 const EntityLayoutWrapper = (props: { children?: ReactNode }) => {
@@ -342,18 +340,6 @@ const overviewContent = (
 
     <Grid item md={6} xs={12}>
       <EntityCatalogGraphCard variant="gridItem" height={400} />
-    </Grid>
-
-    <Grid item md={6} xs={12}>
-      <GitHubIssuesCard
-        filterBy={{
-          labels: ['bug', 'enhancement'],
-        }}
-        orderBy={{
-          field: 'COMMENTS',
-          direction: 'ASC',
-        }}
-      />
     </Grid>
 
     <EntitySwitch>
