--- query.graphql
@@ -1,77 +1,0 @@
-/* http://localhost:7007/api/auth/github /*
-
-{
-  repository(owner: "backstage", name: "backstage") {
-    issues(first: 1, filterBy: { labels:  ["bug"] }) {
-      nodes {
-        title
-        body
-        bodyHTML
-        bodyText
-        number
-        labels(first: 100) {
-          nodes {
-            color
-            name
-            id
-          }
-        }
-        author {
-          url
-          avatarUrl
-          login
-        }
-      }
-    }
-  }
-}
-
-
-    
-fragment issues on Repository {
-  issues(
-    states: OPEN
-    first: 40
-    orderBy: { field: UPDATED_AT, direction: DESC }
-  ) {
-    totalCount
-    edges {
-      node {
-        assignees(first: 10) {
-          edges {
-            node {
-              avatarUrl
-              login
-            }
-          }
-        }
-        author {
-          login
-        }
-        repository {
-          nameWithOwner
-        }
-        title
-        url
-        participants {
-          totalCount
-        }
-        updatedAt
-        createdAt
-        comments(last: 1) {
-          totalCount
-        }
-      }
-    }
-  }
-}
-
-
-query {
-  
-    backstage: repository(name: "backstage", owner: "backstage") {
-      ...issues
-    }
-  
-}    
-
